+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Patch] Update empty state SVGs to point to new stroked versions in /static/x/empty-states/...
+
 ## 27.1.1 - 2017-10-10
 - [Feature] Ignore bundle from static dir on docs/stoybook.